Added -w, -f, -b, -s flags
Custom Workspace
Force reinstall
Rebuild system Go paths
Special single-user install

+++ install_go.sh
@@ -24,7 +24,16 @@
 #
 #	-p Path for Go binaries
 #
+#   -w Path for workspace
+#
 #	-x Uninstall Go
+#
+#   -b Rebuild system path variables (use the same flags for a normal install but add this)
+#
+#   -s Single user install (install to $HOME/.go instead of /usr/local)
+#
+#   -f Force reinstall of Go. Uninstalls any versions given same install paramenters
+#      and installs from scratch
 
 # Script variables
 #
@@ -32,114 +41,178 @@ WGET=`which wget`
 GOVER="1.3.2"
 GOOS="linux"
 GOARCH="amd64"
-GOCODEDIR="${HOME}/gocode"
+GOCODEDIR="${HOME}/go"
 GODIR="/usr/local"
 UPDATE=false
+SINGLE=false
+FORCE=false
+REBUILD=false
 
 # Script functions
 #
 usage() {
-	echo "Usage: $0 [-u] [-x] [-a <string>] [-r <string>] [-p <string>]" 1>&2
+	echo "Usage: $0 [-u] [-b] [-s] [-f] [-x] [-a <string>] [-w <string>] [-r <string>] [-p <string>]" 1>&2
 	echo ""
 	exit 1
+}
+
+removePaths() {
+    echo "---Deleting paths"
+    sed -i '/# GoLang Paths/d' $HOME/.bashrc
+    sed -i '/export GOROOT/d' $HOME/.bashrc
+    sed -i '/export GOPATH/d' $HOME/.bashrc
+    sed -i '/:$GOPATH/d' $HOME/.bashrc
+    sed -i '/export GOBIN/d' $HOME/.bashrc
+    sed -i '/export GOSRC/d' $HOME/.bashrc
 }
 
 uninstall() {
 	echo "Removing Go"
 
+    if [[ $GOROOT == "" ]]; then
+        echo "ERROR: Can't remove Go, GOROOT not defined"
+        exit 1
+    fi
+
 	echo "Deleting install directory"
 	if [[ -d "$GOROOT" ]]; then
 		sudo rm -r $GOROOT
 	fi
-
-	echo "Deleting paths"
-	sed -i '/# GoLang Paths/d' $HOME/.bashrc
-	sed -i '/export GOROOT/d' $HOME/.bashrc
-	sed -i '/export GOPATH/d' $HOME/.bashrc
-	sed -i '/:$GOROOT/d' $HOME/.bashrc
-	sed -i '/export GOBIN/d' $HOME/.bashrc
-	sed -i '/export GOSRC/d' $HOME/.bashrc
-
+    removePaths
 	echo "Done"
-	exit 0
 }
 
 # Parse arguments
 #
-while getopts ":a:uxr:p:" o; do
+while getopts ":ubsfxa:w:r:p:" o; do
     case "${o}" in
-    	a)
+    	a) # Architecture
 			if ${OPTARG} == "32"; then
 				GOARCH="386"
 			fi
 			;;
-        u)
+        b) # Rebuild paths
+            REBUILD=true
+            ;;
+        w) # Workspace dir
+            GOCODEDIR=${OPTARG}
+            ;;
+        u) # Update to $GOVER
             UPDATE=true
             ;;
-        r)
+        r) # Set version to install
             GOVER=${OPTARG}
             ;;
-        p)
+        p) # Install dir
 			GODIR=${OPTARG}
 			;;
-		x)
+		x) # Uninstall
 			uninstall
+            exit 0
 			;;
-        *)
+        s) # Single-user install
+            GODIR="$HOME/.go"
+            SINGLE=true
+            ;;
+        f) # Force reinstall
+            FORCE=true
+            ;;
+        *) # Other
             usage
             ;;
     esac
 done
 
+# Full Go install path
+#
+GOROOTDIR="${GODIR}/go"
+
+# Check if install directory exists
+#
+if [[ -d "${GODIR}" || -L "${GODIR}" ]] && ! $FORCE && ! $REBUILD && ! $UPDATE; then
+    echo "ERROR: Go appears to already be installed, to force a reinstall use -f"
+    exit 1
+fi
+
+# Forced install message
+#
+if $FORCE; then
+    echo ""
+    echo "## Performing forced reinstall ##"
+    echo ""
+fi
+
+# Rebuild path message
+#
+if $REBUILD; then
+    echo ""
+    echo "## Performing path rebuild ##"
+    echo ""
+fi
+
 # Download archive
 #
 echo "--Fetching go language archive--"
 echo ""
-if ! $WGET "https://storage.googleapis.com/golang/go${GOVER}.${GOOS}-${GOARCH}.tar.gz"; then
-	echo "ERROR: Download failed. Exiting."
-	exit 1
+# Delete any existing file with the same name
+if [[ -f "go${GOVER}.${GOOS}-${GOARCH}.tar.gz" ]]; then
+    rm go${GOVER}.${GOOS}-${GOARCH}.tar.gz
+fi
+if ! $REBUILD; then
+    if ! $WGET "https://storage.googleapis.com/golang/go${GOVER}.${GOOS}-${GOARCH}.tar.gz"; then
+    	echo "ERROR: Download failed. Exiting."
+    	exit 1
+    fi
 fi
 
-# If updating Go, delete old version
+# If updating or force reinstalling Go, delete old version
 #
-if $UPDATE; then
+if $UPDATE || $FORCE; then
 	echo "--Removing old version of Go--"
-	if [[ -d "$GOROOT" ]]; then
-		sudo rm -r $GOROOT
-	fi
+    uninstall
 fi
 
 # Extract Go
 #
-echo "--Extracting Go archive--"
-if [[ ! -d "${GODIR}" || ! -L "${GODIR}" ]]; then
-    mkdir -p ${GODIR}
+if ! $REBUILD; then
+    echo "--Extracting Go archive--"
+    if [[ ! -d "${GODIR}" || ! -L "${GODIR}" ]]; then
+        mkdir -p ${GODIR}
+    fi
+
+    if $SINGLE; then
+        if ! tar -C ${GODIR} -xzf "go${GOVER}.${GOOS}-${GOARCH}.tar.gz"; then
+            exit 1
+        fi
+    else
+        if ! sudo tar -C ${GODIR} -xzf "go${GOVER}.${GOOS}-${GOARCH}.tar.gz"; then
+        	exit 1
+        fi
+    fi
 fi
 
-if ! sudo tar -C ${GODIR} -xzf "go${GOVER}.${GOOS}-${GOARCH}.tar.gz"; then
-	exit 1
-fi
-
-# Add gocode dir
+# Create Go workspace
 #
 if [[ -d "${GOCODEDIR}" && ! -L "${GOCODEDIR}" ]]; then
-    echo "--GoCode folder already made--"
+    echo "--Go Workspace already made--"
 else
-    echo "--Making GoCode folder--"
-    mkdir -p ${GOCODEDIR}/src; echo "---Created go code src folder"
-    mkdir -p ${GOCODEDIR}/bin; echo "---Created go code bin folder"
-    mkdir -p ${GOCODEDIR}/pkg; echo "---Created go code pkg folder"
+    echo "--Making Go Workspace--"
+    mkdir -p ${GOCODEDIR}/src; echo "---Created workspace src folder"
+    mkdir -p ${GOCODEDIR}/bin; echo "---Created workspace bin folder"
+    mkdir -p ${GOCODEDIR}/pkg; echo "---Created workspace pkg folder"
 fi
 
 # Add variables to bashrc
 #
 echo "--Writting Go variables--"
 touch $HOME/.bashrc
-if grep -q "# GoLang Paths" "${HOME}/.bashrc"; then
+if grep -q "# GoLang Paths" "${HOME}/.bashrc" && ! $REBUILD; then
 	echo "---Go bin already added to PATH"
 else
+    removePaths
+    echo "---Adding Paths"
 	echo '# GoLang Paths' >> $HOME/.bashrc
-	echo "export GOROOT=${GODIR}/go" >> $HOME/.bashrc
+	echo "export GOROOT=${GOROOTDIR}" >> $HOME/.bashrc
 	echo "export GOPATH=${GOCODEDIR}" >> $HOME/.bashrc
 	echo 'export PATH=$PATH:$GOPATH/bin:$GOROOT/bin' >> $HOME/.bashrc
 	echo 'export GOBIN=$GOPATH/bin' >> $HOME/.bashrc
@@ -150,9 +223,13 @@ fi
 #
 echo "--Cleaning up--"
 source $HOME/.bashrc
-rm go${GOVER}.${GOOS}-${GOARCH}.tar.gz
+if [[ -f "go${GOVER}.${GOOS}-${GOARCH}.tar.gz" ]]; then
+    rm go${GOVER}.${GOOS}-${GOARCH}.tar.gz
+fi
 echo "---Successfully installed Go ${GOVER}"
 
+# Thank you message
+#
 cat <<EOF
 
   Thank you for installing Go!
@@ -164,5 +241,10 @@ cat <<EOF
   You can see all the go command line options by
   running the following command:
      go help
+
+  This script was created by DragonRider23.
+  https://github.com/dragonrider23/go-install-tool
+
+  Released under the MIT license.
   
 EOF
